Move the cli modules
